misc: import i18n messages as JSON modules
As far as I can tell, import attribute syntax should work everywhere
lighthouse is supported.

My intent behind doing this was to maybe get Lighthouse to the point
where you can bundle it for NodeJS using something like esbuild or
rollup. But it turns out that errors about these locales files were just
the first ones to occur when executing a bundled lighthouse entrypoint
:^). Audits and gatherers are still loaded completely dynamically from
the filesystem. Maybe those could be dynamically imported instead, but
that would be a more substantial change.

Anyway, perhaps it's still better to ship this than to not.

Fixes #14611.