fix(viewer): only disable save-gist button on successful save
Previously, the "Save as Gist" button was disabled immediately after any save attempt,
even if the gist creation failed (for example, if the user was not logged in).

This change updates the logic so that the button is disabled only after a confirmed,
successful save (when a gist ID is returned). This allows users to retry saving in case of failure.

Fixes #14947